+++ samples/lib/widgets/pagination.dart
@@ -42,7 +42,7 @@ class _PaginationPageState extends StatefulSampleState<PaginationPage> {
   @override
   Widget sample(BuildContext context) => Column(
     mainAxisAlignment: .center,
-    children: [FPagination(controller: _controller)],
+    children: [FPagination(control: .managed(controller: _controller))],
   );
 }
 
@@ -70,8 +70,7 @@ class _PaginationCustomIconPageState extends StatefulSampleState<PaginationCusto
       mainAxisAlignment: .center,
       children: [
         FPagination(
-          controller: _controller,
-          next: Padding(
+          control: .managed(controller: _controller), next: Padding(
             padding: style.itemPadding,
             child: ConstrainedBox(
               constraints: style.itemConstraints,
@@ -174,7 +173,7 @@ class _PaginationWithViewPageState extends StatefulSampleState<PaginationWithVie
             ),
           ),
         ),
-        FPagination(controller: _paginationController, onChange: _handlePageChange),
+        FPagination(control: .managed(controller: _paginationController, onChange: _handlePageChange)),
       ],
     );
   }